+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [0.6.1]
+- optimize error report for pratt parser
+- optimize depth for pratt parser.
+- allow |_ to chaincall. replace arg of first call with _
+- allow chaincall as alias.
+
 ## [0.6.0]
 **broken changes**
 - change `?!` to `?>` : return err map as result.